mevbot/strategies/sandwich.py
"""Sandwich attack strategy implementation"""
from typing import Dict, List, Optional, Tuple
from web3 import Web3
from eth_typing import Address
import asyncio
import logging
from dataclasses import dataclass
import numpy as np
from ..core.simd import SIMDProcessor
from ..core.flashbots import FlashbotsManager

logger = logging.getLogger(__name__)

@dataclass
class SandwichParams:
    target_tx_hash: str
    token: Address
    pool: Address
    front_amount: int
    back_amount: int
    expected_profit: float
    front_tx: Optional[Dict] = None
    back_tx: Optional[Dict] = None

class SandwichStrategy:
    def __init__(self, w3: Web3, flashbots: FlashbotsManager, config: Dict):
        self.w3 = w3
        self.flashbots = flashbots
        self.config = config
        self.simd = SIMDProcessor()
        
        # Performance tracking
        self.total_attempts = 0
        self.successful_attempts = 0
        self.total_profit = 0.0
        self.total_gas_used = 0
        
    async def analyze_opportunity(
        self,
        target_tx: Dict,
        pool_info: Dict
    ) -> Optional[SandwichParams]:
        """Analyze if a transaction is suitable for sandwich attack"""
        try:
            # Extract parameters
            token = pool_info['token']
            pool = pool_info['address']
            target_amount = int(target_tx.get('value', 0))
            liquidity = await self._get_pool_liquidity(pool)
            gas_price = await self.flashbots.get_gas_price()
            
            if target_amount == 0 or liquidity == 0:
                return None
                
            # Use SIMD to calculate optimal amounts
            amounts, profits = self.simd.batch_sandwich_optimization(
                [float(target_amount)],
                [float(liquidity)],
                [float(gas_price)]
            )
            
            front_amount = int(amounts[0])
            expected_profit = float(profits[0])
            
            if front_amount == 0 or expected_profit <= 0:
                return None
                
            # Calculate backrun amount (usually slightly higher than frontrun)
            back_amount = int(front_amount * 1.02)  # Add 2% to ensure success
            
            return SandwichParams(
                target_tx_hash=target_tx['hash'],
                token=token,
                pool=pool,
                front_amount=front_amount,
                back_amount=back_amount,
                expected_profit=expected_profit
            )
            
        except Exception as e:
            logger.error(f"Error analyzing sandwich opportunity: {e}")
            return None
            
    async def execute_sandwich(self, params: SandwichParams) -> bool:
        """Execute a sandwich attack"""
        try:
            # Create frontrun transaction
            params.front_tx = await self._create_swap_tx(
                params.token,
                params.pool,
                params.front_amount,
                True  # is_buy
            )
            
            # Create backrun transaction
            params.back_tx = await self._create_swap_tx(
                params.token,
                params.pool,
                params.back_amount,
                False  # is_sell
            )
            
            # Submit bundle to Flashbots
            bundle = [
                params.front_tx,
                {'hash': params.target_tx_hash},  # Target transaction
                params.back_tx
            ]
            
            success = await self.flashbots.send_bundle(bundle)
            
            # Update statistics
            self.total_attempts += 1
            if success:
                self.successful_attempts += 1
                self.total_profit += params.expected_profit
                self.total_gas_used += (
                    params.front_tx.get('gas', 0) +
                    params.back_tx.get('gas', 0)
                )
                
            return success
            
        except Exception as e:
            logger.error(f"Error executing sandwich attack: {e}")
            return False
            
    async def _get_pool_liquidity(self, pool: Address) -> int:
        """Get current liquidity of a pool"""
        # This would call the actual pool contract
        # For now, return a dummy value
        return int(1e20)
        
    async def _create_swap_tx(
        self,
        token: Address,
        pool: Address,
        amount: int,
        is_buy: bool
    ) -> Dict:
        """Create a swap transaction"""
        # This would create the actual swap transaction
        # For now, return a dummy transaction
        return {
            'to': pool,
            'value': amount if is_buy else 0,
            'gas': 150000,
            'gasPrice': await self.flashbots.get_gas_price()
        }
